Move `DandisetOwnersDialog` to `components`
